pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-kg"
version = "0.1.0"
description = "Multi-agent knowledge graph construction from business documents"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "instructor>=1.0",
    "openai>=1.0",
    "agent-framework>=0.1.0a0",
    "tiktoken",
    "numpy",
    "scikit-learn",
    "neo4j>=5.0",
]

[project.optional-dependencies]
hdbscan = ["hdbscan", "umap-learn"]
dev = ["pytest", "pytest-asyncio", "ruff"]

[tool.hatch.build.targets.wheel]
packages = ["src/agent_kg"]

[tool.ruff]
line-length = 100
target-version = "py311"
